- creato gestione appartamenti e tenant

+++ core-service/src/main/java/com/propertypilot/coreservice/repository/AppartamentoTenantRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface AppartamentoTenantRepository extends JpaRepository<AppartamentoTenant, Long> {
-    Optional<AppartamentoTenant> findByAppartamentoId(Long appartamentoId);
+    Optional<AppartamentoTenant> findByAppartamentoId(Integer appartamentoId);
 }
